Make TCP logstash work (#284)
* Make TCP logstash work

Currently, logstash.LogstashHandler is an alias for UDPLogstashHandler - thus, handler TCP is ignored, and glastopf always uses UDP logstash.  This change makes it work properly.

* Simplify logstash ingest w/log components split

Python logstash has a feature that lets you supply extra options to the
logger.  These extra options are in the format of a dictionary, and
those get passed along to logstash and the rest of the chain.

This patch breaks out the things that form the message, and supplies
those as extra options.  Without this - logstash users have to use a
grok expression to split things out, and that's silly.